refactor: integrations auth (#489)
* refactor: refactor auth for crm integrations

* fix: fix wrong imports

* fix: remove unnecessary imports

* refactor: refactor auth for chat integrations

* refactor: refactor auth for ticket integrations

* rename utils directory

* refactor abstracted error handler to handle success and error outcomes

* use outcome handler in chat authHandlers

* use outcome handler in crm authHandlers

* use outcome handler in ticket authHandlers

* use outcome handler in main auth file

* use outcome handler in main auth file

* refactor: add `baseOAuthHandler` and convert all handlers to extend this class
