+++ packages/backend/helpers/auth/baseOAuthHandler.ts
@@ -1,0 +1,10 @@
+import { IntegrationAuthProps } from '../../constants/common';
+import { Response } from 'express';
+
+abstract class BaseOAuthHandler {
+    async handleOAuth(_integrationAuthProps: IntegrationAuthProps): Promise<Response<any, Record<string, any>> | void> {
+        return;
+    }
+}
+
+export default BaseOAuthHandler;
